* Fixes size display for video files larger than 2GB.
git-svn-id: https://burn-osx.svn.sourceforge.net/svnroot/burn-osx/trunk@373 df5e183e-a46d-47f3-acf2-86dc86df8d71

+++ burn/Source/KWVideoController.m
@@ -197,7 +197,7 @@
 					[rowData setObject:[NSArray array] forKey:@"Chapters"];
 				}
 				
-				int displaySize = [[attrib objectForKey:NSFileSize] floatValue];
+				float displaySize = [[attrib objectForKey:NSFileSize] floatValue];
 				
 					if (selrow < 2)
 					{
